fixed all conversations

+++ src/main/java/org/example/librarymanagement/service/implement/BookFileProcess.java
@@ -14,6 +14,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.Reader;
+import java.math.BigDecimal;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -72,7 +73,7 @@ public class BookFileProcess {
                 book.setTitle(title);
                 book.setDescription(description);
                 book.setQuantity(quantity);
-                book.setPrice(price);
+                book.setPrice(BigDecimal.valueOf(price));
                 book.setGenreId(genreId);
                 book.setAuthorId(authorId);
                 bookList.add(book);
